back/requirements.txt
altgraph==0.17.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==25.1.0
babel==2.17.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.9.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.16
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.8.0
dxcam==0.0.5
executing==2.2.0
fastapi==0.135.1
fastjsonschema==2.21.1
filelock==3.17.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
fsspec==2025.3.0
future==1.0.0
greenlet==3.3.2
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
humanfriendly==10.0
idna==3.7
Imagegrab==0.0.3
interception-python==0.0.1
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
iso8601==2.1.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keyboard==0.13.5
kiwisolver==1.4.8
lap==0.5.12
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mistune==3.1.1
motor==3.7.1
MouseInfo==0.1.3
mpmath==1.3.0
mss==9.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.26.4
onnx==1.17.0
onnxruntime==1.21.0
onnxruntime-directml==1.21.0
onnxslim==0.1.49
opencv-python==4.9.0.80
overrides==7.7.0
packaging==24.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.3.6
prometheus_client==0.21.1
prompt_toolkit==3.0.50
protobuf==6.30.2
psutil==6.1.1
psycopg2-binary==2.9.11
pure_eval==0.2.3
py-cpuinfo==9.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
Pymem==1.14.0
pymongo==4.16.0
PyMsgBox==1.0.9
pynput==1.7.6
pyparsing==3.2.1
pyperclip==1.8.2
pyreadline3==3.5.4
PyRect==0.2.0
PyScreeze==0.1.30
passlib==1.7.4
pyotp==2.9.0
PyJWT==2.8.0
pyserial==3.5
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-json-logger==3.2.1
python-telegram-bot==20.6
pytweening==1.2.0
pytz==2025.1
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
schedule==1.2.2
scipy==1.15.2
screeninfo==0.8.1
seaborn==0.13.2
Send2Trash==1.8.3
serial==0.0.97
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
spark-parser==1.8.9
SQLAlchemy==2.0.48
pytest==8.3.4
pytest-asyncio==0.24.0
stack-data==0.6.3
starlette==0.52.1
sympy==1.13.1
telegram-send==0.37
terminado==0.18.1
tinycss2==1.4.0
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
ultralytics==8.3.87
ultralytics-thop==2.0.14
uncompyle6==3.9.2
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.42.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
prometheus-fastapi-instrumentator==7.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
windows-capture==1.5.0
xdis==6.1.3
